+++ src/rdtime.h
@@ -174,7 +174,7 @@ static RD_INLINE int rd_timeout_remains (rd_ts_t abs_timeout) {
             timeout_us == RD_POLL_NOWAIT)
                 return (int)timeout_us;
 
-        return timeout_us / 1000;
+        return (int)(timeout_us / 1000);
 }
 
 /**
